annotation: autosave comments when lose focus without exiting editing
how it is expected to work:
before this patch, whenever comment is saved,
editing comment box(DOM elements) is deleted and
a new comment(DOM elements) is created when we recieve the data from server about saved comment

This commit prevents the deletion of the original comment box which is being edited,
if comment loses focus it is saved. If user explicitly saves the comment then it
follows the routine which were used previously.

Many checks were added to prevent comments from getting unselected(select means it is highliting the text in doc)
or getting collapsed

Adapt patch to LibreOffice Online.

Signed-off-by: Pranam Lashkari <lpranam@collabora.com>
Change-Id: If24ccbcd7a0287b659f790db486ada5d17608a32