wsd: test: exclude readonly sessions from modification indication
New unit-test that reproduces the corner-case
with disconnection. With graceful unloading
a previous unit-test verified that it worked
as expected.

This is restored from e6798e9062fbb1e2466afdfd094a767e572fb4d5
after it was reverted (due to a PDF comment saving regression).
Since the issue is fixed separately in
0b27e85fc56f41463d24e041057709182db8532c, these tests can
now be restored.

Change-Id: Id84cda4f4599c559018247c32ea1205e154e4984
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>