different 'parts' in requestedTiles on switching calc tabs
if we switch calc tabs quickly that can result in still unfulfilled
requests for the first tab in requestedTiles when the new requests are
compared with the existing one for duplicates.

It doesn't make sense to replace a request for a new tile with an old
request from a different part. But tilecombines are intended to share a
'part', so a single tilecombine shouldn't be synthesized from multiple
requests for different parts.

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: Id8d40f0a7e7f9d32fed67a415756284d7f6a53d4