Copying the hyperlink issue: * Add a paragraph. * Select the whole paragraph and add comment. * Add hyperlink to a word in the paragraph. * Try to copy the hyperlink via its menu. * It sometimes doesn't copy the hyperlink.
We are checking if a comment has focus or not before doing the copy operation. This check is for another issue.
I changed the check with "app.view.commentHasFocus" which seems reliable.
Then tested the previous fix.

Also commented out code is removed.
White space change is automatic.

Adapt patch to LibreOffice Online.

Signed-off-by: Gökay Şatır <gokaysatir@gmail.com>
Change-Id: Ibc2c5e550eb8a3e24b734e0371c87dac44b2ff6d